# Conflicts:
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/Channels/UpdateChannelMembersPayload.cs.meta
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/ChatController.asmdef
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/ChatController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/ChatMessage/ChatMessage.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/ChatMessage/ChatMessageListPayload.cs.meta
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/IChatController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ChatController/Tests/Helpers/ChatController_Mock.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/ExploreV2/Scripts/MainMenu/ExploreV2Menu/ExploreV2MenuComponentController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/FriendsController/LazyLoading/LazyLoadingFriendsControllerMock.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/ChatWidgetHUD/ChatHUDView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/FriendsHUD/Scripts/FriendsHUDController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/FriendsHUD/Scripts/Tabs/FriendRequestsTabComponentView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/FriendsHUD/Tests/FriendsHUDControllerShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/HUDController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/HUDFactory.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NotificationBadge/ReceivedFriendRequestsNotificationBadge.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NotificationBadge/Tests/UnreadNotificationBadgeShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NotificationBadge/UnreadNotificationBadge.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/NotificationBadge/UnreadWorldNotificationBadge.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PrivateChatWindow/PrivateChatWindowComponentView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PrivateChatWindow/PrivateChatWindowController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/PrivateChatWindow/Tests/PrivateChatWindowControllerShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/SocialBarPrefabs/Resources/SocialBarV1/ConversationListHUD.prefab
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/SocialBarPrefabs/Resources/SocialBarV1/FriendsHUD.prefab
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/SocialBarPrefabs/Resources/SocialBarV1/PrivateChatHUD.prefab
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/SocialBarPrefabs/UserContextMenu/UserContextMenuPanel.prefab
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/TaskbarHUD/Resources/Taskbar.prefab
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/TaskbarHUD/Tests/TaskbarHUDShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/TaskbarHUD/Tests/WorldChatWindowViewMock.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/CollapsableChatSearchListComponentView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/CollapsablePublicChannelListComponentView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/IWorldChatWindowView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/PrivateChatEntry.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/PublicChannelEntry.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/PublicChatWindowController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/Tests/PublicChannelEntryShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/Tests/WorldChatWindowComponentViewShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/Tests/WorldChatWindowControllerShould.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/WorldChatWindowComponentView.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/HUD/WorldChatWindowHUD/WorldChatWindowController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/Controllers/PluginSystem/PluginSystemFactory.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/UserProfile/UserProfileController.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/WebInterface/Interface.cs
#	unity-renderer/Assets/Scripts/MainScripts/DCL/WorldRuntime/KernelCommunication/WebSocketCommunication/WebSocketCommunication.cs

+++ unity-renderer/Assets/DCLPlugins/ECS7/TestsUtils/Components/UI/ECS7TestUIUtils.cs
@@ -8,7 +8,8 @@ namespace DCLPlugins.ECS7.TestsUtils
         public static PBUiTransform CreatePBUiTransformDefaultModel()
         {
             var model = new PBUiTransform();
-            model.Parent = SpecialEntityId.SCENE_ROOT_ENTITY;
+            // EntityId long to int check
+            model.Parent = (int)SpecialEntityId.SCENE_ROOT_ENTITY;
             model.AlignContent = YGAlign.FlexStart;
             model.AlignSelf = YGAlign.Auto;
             model.AlignItems = YGAlign.Auto;
